fix!: refactor project structure to fix config
Despite overriding the app config for tests, the tests would still
commit to the production db. The reason for this was because there was
no package holding together all the modules, so everything was imported
relatively. To fix this, it is now backend/src to avoid naming
conflicts, and test folder was also moved up. A side effect of this is
now the pytest command works normally with no need to call it as a
module.
