--- backend/app/.env
@@ -1,2 +1,0 @@
-MONGO_URI="mongodb://127.0.0.1:27017"
-SECRET_KEY=secret
